Fixed a few issues, added support for more symbols.
Fixed support for "<=" and ">=" in PopcornParser (order in RelExpr).
Fixed support for ">" in Expressions (was same as ">=").
Changed interval1#interval to return List instead of Stream.
Added support for prog1#if, logic1#not, list1#map, list3#entry, nums1#e,
  nums1#NaN and nums1#isNaN (which is not actually defined in nums1).
Support shorthand versions of if/for/while.
Modified and simplified setting of POJO properties.